fix: error messages are ignored when executing resource nodes
